generator/requirements.txt
sqlalchemy
PyMySQL
python-dotenv
Faker
boto3